Move LegacyASTTransformer and SyntaxASTMapping to AST
